back to ll

+++ math/combinatorics.cpp
@@ -15,13 +15,13 @@
 */
 
 const int ms=10000;
-vector<vector<long>> c;
+vector<vector<long long>> c;
 
 void initcomb(){
     c.resize(ms+1,vector<ll>(ms/2+1,0));
 }
 
-long comb(int n,int k){//combinacao de n, escolhendo k
+long long comb(int n,int k){//combinacao de n, escolhendo k
     if(k==0 or k==n) return 1;
     k = min(k,n-k);
     if(c[n][k]) return c[n][k];
