+++ math/eulerfunction.cpp
@@ -1,11 +1,11 @@
 
 const int P = 2e7+5;
-vector<long> phi(P+1);
+vector<long long> phi(P+1);
 void PHI(){//get prime factors and calculate phi
-	for(long i=1;i<P;i++)phi[i]=i;
-	for(long i=2;i<P;i++){//O(N)
+	for(long long i=1;i<P;i++)phi[i]=i;
+	for(long long i=2;i<P;i++){//O(N)
 		if(phi[i]==i){//n is prime
-			for(long j=i;j<P;j+=i){//O(logi)
+			for(long long j=i;j<P;j+=i){//O(logi)
 				//i is a prime factor of j
 				phi[j]=phi[j]/i*(i-1);
 			}
@@ -13,9 +13,9 @@ void PHI(){//get prime factors and calculate phi
 	}
 }
 
-long PHI(long n){//calculates one phi
-    long ans = n;
-	for(long i=2;i*i<=n;i++){
+long long PHI(long long n){//calculates one phi
+    long long ans = n;
+	for(long long i=2;i*i<=n;i++){
 		if(n%i==0){
 			ans-=ans/i;
 			while(n%i==0)n/=i;
