+++ math/pollard-rho-miler-rabin.cpp
@@ -1,16 +1,16 @@
 
 //requires fexp.cpp
-bool overflow(long a, long b) {
-  return b && (a >= (1long << 62) / b);
+bool overflow(long long a, long long b) {
+  return b && (a >= (1long long << 62) / b);
 }
 
-long add(long a, long b, long mod) {
+long long add(long long a, long long b, long long mod) {
   return (a + b) % mod;
 }
 
-long mul(long a, long b, long mod) {
+long long mul(long long a, long long b, long long mod) {
   if (!overflow(a, b)) return (a * b) % mod;
-  long ans = 0;
+  long long ans = 0;
   while(b) {
     if (b & 1) ans = add(ans, a, mod);
     a = add(a, a, mod);
@@ -19,20 +19,20 @@ long mul(long a, long b, long mod) {
   return ans;
 }
 
-long my_rand() {
-  long ans = rand();
+long long my_rand() {
+  long long ans = rand();
   ans = (ans << 31) | rand();
   return ans;
 }
 
-bool miller(long p, int precision) {//check primality
+bool miller(long long p, int precision) {//check primality
   if(p < 2) return 0;
   if(p % 2 == 0) return (p == 2);
-  long s = p - 1;
+  long long s = p - 1;
   while(s % 2 == 0) s >>= 1;
   for(int i = 0; i < precision; i++) {
-    long a = rand() % (p - 1) + 1, temp = s;
-    long mod = fexp(a, temp, p);
+    long long a = rand() % (p - 1) + 1, temp = s;
+    long long mod = fexp(a, temp, p);
     while(temp != p - 1 && mod != 1 && mod != p - 1) {
       mod = mul(mod, mod, p);
       temp <<= 1;
@@ -42,12 +42,12 @@ bool miller(long p, int precision) {//check primality
   return 1;
 }
 
-long rho(long n) {//find a prime factor
+long long rho(long long n) {//find a prime factor
   if (n == 1 || miller(n, 10)) return n;
   if (n % 2 == 0) return 2;
   while(1) {
-    long x = my_rand() % (n - 2) + 2, y = x;
-    long c = 0, cur = 1;
+    long long x = my_rand() % (n - 2) + 2, y = x;
+    long long c = 0, cur = 1;
     while(c == 0) {
       c = my_rand() % (n - 2) + 1;
     }
